决定自己写项目,而不是依赖视屏内容,我也不知道能不能把这个项目坚持到上线...

+++ src/main/java/com/mralmost/community/controller/PublishController.java
@@ -47,7 +47,7 @@ public class PublishController {
             @RequestParam(name = "title", required = false) String title,
             @RequestParam(name = "description", required = false) String description,
             @RequestParam(name = "tag", required = false) String tag,
-            @RequestParam(name = "id", required = false) Integer id,
+            @RequestParam(name = "id", required = false) Long id,
             HttpServletRequest request,
             Model model) {
 
@@ -91,7 +91,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model) {
         QuestionDTO byId = questionService.findById(id);
         model.addAttribute("title", byId.getTitle());
